+++ js/lite/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @gradio/lite
 
+## 4.21.0
+
+### Features
+
+- [#7647](https://github.com/gradio-app/gradio/pull/7647) [`57510f9`](https://github.com/gradio-app/gradio/commit/57510f9ce0b38d473cc90800a05e769f1a47be9b) - Lite version.  Thanks @pngwn!
+
+### Dependency updates
+
+- @gradio/wasm@0.8.0
+- gradio@4.21.0
+
 ## 4.14.2
 
 ### Features